device-service/pyproject.toml
[project]
name = "device-management-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic",
    "asyncpg",
    "fastapi",
    "pydantic",
    "pydantic-settings",
    "python-jose[cryptography]",
    "sqlalchemy[asyncio]",
    "uvicorn",
]

[project.scripts]
dev = "app.cli:dev"

[build-system]
requires = ["uv-build"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
module-name = "app"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
